simple/python/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
kiteconnect==4.2.0
itsdangerous==2.2.0
